+++ nestkernel/histentry.h
@@ -35,9 +35,9 @@ namespace nest
 class histentry
 {
 public:
-  histentry( double t, double Kminus, double Kminus_triplet, size_t access_counter );
+  histentry( size_t t, double Kminus, double Kminus_triplet, size_t access_counter );
 
-  double t_;              //!< point in time when spike occurred (in ms)
+  size_t t_;              //!< point in time when spike occurred (in ms)
   double Kminus_;         //!< value of Kminus at that time
   double Kminus_triplet_; //!< value of triplet STDP Kminus at that time
   size_t access_counter_; //!< access counter to enable removal of the entry, once all neurons read it
